refactor: 调整部分页面布局，统一卡片样式 feat: 新增 BeautifulCard 卡片组件 style: 统一 RunnerScreen 卡片样式
