+++ app/src/main/java/yangfentuozi/runner/app/ui/activity/envmanage/EnvManageActivity.kt
@@ -3,7 +3,7 @@ package yangfentuozi.runner.app.ui.activity.envmanage
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import yangfentuozi.runner.app.base.BaseActivity
-import yangfentuozi.runner.app.ui.screens.EnvManageScreen
+import yangfentuozi.runner.app.ui.screens.envmanager.EnvManageScreen
 import yangfentuozi.runner.app.ui.theme.RunnerTheme
 
 class EnvManageActivity : BaseActivity() {
